reset_window_pos: a new plugin,to reset dock positions.
reset_window_pos.py is a new plugin to reset
dock positions of main window.

This plugin records the startup positions of each docks,
and when activated, set their positions as startup.

This is created to deal with annoying automatic resize of main window,
which occasionary occur in some version of Gtk3.